requirements_fastapi.txt
# FastAPI Requirements - Optimized for Performance
# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.7

# Data validation and serialization
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# Optional: Development and monitoring
watchfiles==0.21.0  # For auto-reload in development

# Security and Authentication
python-multipart==0.0.6  # For form data handling
python-jose[cryptography]==3.3.0  # For JWT tokens
authlib==1.2.1  # For OAuth2 integration
httpx==0.25.2  # For async HTTP requests (OAuth)
itsdangerous==2.1.2  # For secure session management

# Monitoring and logging (optional)
structlog==23.2.0

# Performance monitoring (optional)
prometheus-client==0.19.0 